added a helper class to build the string for the port service calls
