+++ Main/SpringExample/demo-spring-app-master/src/main/java/ServiceRequestor/HttpConnectionHandler.java
@@ -1,8 +1,11 @@
 package ServiceRequestor;
 
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 
 public class HttpConnectionHandler
 {
@@ -27,5 +30,26 @@ public class HttpConnectionHandler
     public static void disconnectHttp(HttpURLConnection connection)
     {
         connection.disconnect();
+    }
+
+    public static void writeToOutputStream(HttpURLConnection conn, String input) throws IOException
+    {
+        var os = conn.getOutputStream();
+        os.write(input.getBytes(StandardCharsets.UTF_8));
+        os.flush();
+    }
+
+    public static String readBuffer(HttpURLConnection conn) throws IOException
+    {
+        BufferedReader br = new BufferedReader(new InputStreamReader(conn.getInputStream()));
+        var output = "";
+        while(br.readLine() != null)
+        {
+            output += br.readLine();
+            //// TODO log output
+            //// TODO handle the output.
+        }
+
+        return output;
     }
 }
